Migrate credentials to Jenkins
