+++ builds/jazz-build-module/service-configuration-data-loader.groovy
@@ -2,7 +2,6 @@
 import groovy.json.JsonSlurperClassic
 import groovy.json.JsonOutput
 import groovy.transform.Field
-import jenkins.security.*
 
 echo "Service configuration module loaded successfully"
 
@@ -334,9 +333,9 @@ def loadServiceConfigurationData() {
 			sh "sed -i -- 's/{job_token}/${config_loader.JENKINS.JOB_AUTH_TOKEN}/g' ./config/stg-config.json"
 			sh "sed -i -- 's/{job_token}/${config_loader.JENKINS.JOB_AUTH_TOKEN}/g' ./config/prod-config.json"
 
-			sh "sed -i -- 's/{api_token}/${getApiToken()}/g' ./config/dev-config.json"
-			sh "sed -i -- 's/{api_token}/${getApiToken()}/g' ./config/stg-config.json"
-			sh "sed -i -- 's/{api_token}/${getApiToken()}/g' ./config/prod-config.json"
+			sh "sed -i -- 's/{api_token}/${utilModule.getApiToken()}/g' ./config/dev-config.json"
+			sh "sed -i -- 's/{api_token}/${utilModule.getApiToken()}/g' ./config/stg-config.json"
+			sh "sed -i -- 's/{api_token}/${utilModule.getApiToken()}/g' ./config/prod-config.json"
 
 			sh "sed -i -- 's/{conf-region}/${region}/g' ./config/dev-config.json"
 			sh "sed -i -- 's/{conf-region}/${region}/g' ./config/stg-config.json"
@@ -399,13 +398,15 @@ def loadServiceConfigurationData() {
 				sh "sed -i -- 's,{bb_service_host},http://${config_loader.REPOSITORY.BASE_URL},g' ./config/stg-config.json"
 				sh "sed -i -- 's,{bb_service_host},http://${config_loader.REPOSITORY.BASE_URL},g' ./config/prod-config.json"
 
-				sh "sed -i -- 's/{bb_username}/${config_loader.SCM.USERNAME}/g' ./config/dev-config.json"
-				sh "sed -i -- 's/{bb_username}/${config_loader.SCM.USERNAME}/g' ./config/stg-config.json"
-				sh "sed -i -- 's/{bb_username}/${config_loader.SCM.USERNAME}/g' ./config/prod-config.json"
+        withCredentials([[$class: 'UsernamePasswordMultiBinding', credentialsId: config_loader.REPOSITORY.CREDENTIAL_ID, passwordVariable: 'PWD', usernameVariable: 'UNAME']]) {
+          sh "sed -i -- 's/{bb_username}/${UNAME}/g' ./config/dev-config.json"
+          sh "sed -i -- 's/{bb_username}/${UNAME}/g' ./config/stg-config.json"
+          sh "sed -i -- 's/{bb_username}/${UNAME}/g' ./config/prod-config.json"
 
-				sh "sed -i -- 's/{bb_password}/${config_loader.SCM.PASSWORD}/g' ./config/dev-config.json"
-				sh "sed -i -- 's/{bb_password}/${config_loader.SCM.PASSWORD}/g' ./config/stg-config.json"
-				sh "sed -i -- 's/{bb_password}/${config_loader.SCM.PASSWORD}/g' ./config/prod-config.json"
+          sh "sed -i -- 's/{bb_password}/${PWD}/g' ./config/dev-config.json"
+          sh "sed -i -- 's/{bb_password}/${PWD}/g' ./config/stg-config.json"
+          sh "sed -i -- 's/{bb_password}/${PWD}/g' ./config/prod-config.json"
+        }
 			}
 
 			if (config_loader.SCM.TYPE == "gitlab") {
@@ -483,15 +484,4 @@ def setLogStreamPermission(config){
 		}
 	}
 }
-
-def getApiToken(){
-	withCredentials([[$class: 'UsernamePasswordMultiBinding', credentialsId: config_loader.JENKINS.CREDENTIAL_ID, passwordVariable: 'PWD', usernameVariable: 'UNAME']]){
-		User u = User.get(UNAME)
-		ApiTokenProperty t = u.getProperty(ApiTokenProperty.class)
-		def token = t.getApiToken()
-		return token
-	}
-}
-
-
 return this
