+++ builds/service-onboarding-build-pack/Jenkinsfile
@@ -168,10 +168,10 @@ node  {
 				// Add service_id  in deployment-env.yml
 
 				sh "echo 'service_id: $service_id' >> ./deployment-env.yml"
-				
+
 				sh "sed -i -- 's/{service_name}/${service_config['service']}/g' ./Jenkinsfile"
 				sh "sed -i -- 's/{domain}/${service_config['domain']}/g' ./Jenkinsfile"
-				
+
 			}
 			events.sendCompletedEvent('MODIFY_TEMPLATE', 'modify template completed', context_map)
 		}
@@ -221,12 +221,14 @@ node  {
 		{
 			sh "ls -lart"
 
-            withCredentials([[$class: 'UsernamePasswordMultiBinding', credentialsId: repo_credential_id, passwordVariable: 'PWD', usernameVariable: 'UNAME']]) {
+        withCredentials([[$class: 'UsernamePasswordMultiBinding', credentialsId: repo_credential_id, passwordVariable: 'PWD', usernameVariable: 'UNAME']]) {
 				try {
 					events.sendStartedEvent("ADD_WEBHOOK", 'service repo creation starts', context_map)
 					def scm_webhook_target_url
 					if (configLoader.SCM.TYPE == "gitlab") {
-						scm_webhook_target_url = "http://${configLoader.JENKINS.USERNAME}:${configLoader.JENKINS.JENKINS_PASSWORD}@${JenkinsLocationConfiguration.get().getUrl().split('/')[2]}/project/Gitlab-Trigger-Job"
+            withCredentials([[$class: 'UsernamePasswordMultiBinding', credentialsId: configLoader.JENKINS.CREDENTIAL_ID, passwordVariable: 'PWD', usernameVariable: 'UNAME']]){
+              scm_webhook_target_url = "http://${UNAME}:${utilModule.getApiToken()}@${JenkinsLocationConfiguration.get().getUrl().split('/')[2]}/project/Gitlab-Trigger-Job?token=${configLoader.JENKINS.JOB_AUTH_TOKEN}"
+            }
 					} else if (configLoader.SCM.TYPE == "bitbucket") {
 						scm_webhook_target_url = JenkinsLocationConfiguration.get().getUrl() + "/bitbucket-scmsource-hook/notify"
 					}
@@ -434,7 +436,7 @@ def loadBuildModules(buildModuleUrl){
 		configModule = load "config-loader.groovy"
 		configLoader = configModule.initialize(resultJsonString)
 		echo "config loader loaded successfully."
-		
+
 		scmModule = load "scm-module.groovy"
 		scmModule.initialize(configLoader)
 		echo "SCM module loaded successfully."
@@ -445,7 +447,7 @@ def loadBuildModules(buildModuleUrl){
 		serviceMetadataLoader = load "service-metadata-loader.groovy"
 		serviceMetadataLoader.initialize(configLoader)
 		echo "Service metadata loader module loaded successfully."
-    
+
 		utilModule = load "utility-loader.groovy"
 	}
 }
